requirements.txt
flask>=1.0.2
requests>=2.20.1
cassandra-driver>=3.16.0